Reorganize mapping strategies.
